Re-organizing code for multiple platforms.
We're going to try some things out and see what happens. Specifically, we're going to set things up to support both the 328-based Arduinos as well as the 1280-based Mega. This is because we're using the Ardupilot Mega, and need support for additional USARTs, etc.

Christian and I had kicked around a directory structure, but I'm not sure we can base this on processor. (Or, we could... but...) My concern is this: not all 1280-based Arduino products are created equal. For example, the Seeed Studio model exposes 70 digital pins, while the "original" Arduino Mega exposes 54. They're both 1280-based devices, but one has a larger ("mega") footprint, while another has a traditional (small, "arduino") footprint. How best to organize these differences? Should we have support files that are ATmega1280 specific, and then another layer that is "board" specific? Or, are they one-and-the-same? 

Hence, we'll start with *some* kind of separation, and can reorganize based on feedback/comments. (Or, drop a note and push things around with us, either way.)