Adding modules to each platform that will hold board-specific pin mappings.
We expect these will contain the PROC "pin.location" as well as the arrays that map pin numbers to registers. Our thought is that higher-level modules (eg. "avr.module") will simply include "pinmapping.module," and the lookup will happen on a board-specific basis.

Someone shout if this is a Bad Direction To Go.