Coppied the pinmapping.module from .../platforms/arduinomega to /ardupilotmega.
As of yet, we've made no changes to this pinmapping.module, dispite expecting it to requre a good number to get working. We have hit a snag getting all 16 adc pins to read though. We suspect the problem to be in the acd PROC in plumbing actually... There seems to be some 328p specific code there.