Adding some examples that test all of the serial ports on the Mega.
What sorta bothers me is that I used interrupt-driven code, but haven't enabled the appropriate interrupt handlers in the firmware. So, somehow, I'm cruising through "wait.for.interrupt" without actually having even mapped the interrupt vectors for the serial ports. That is, we're running a firmware that carries with it interrupt code for the ATmega328p, but it was compiled for the ATmega1280. I did not use the MAP_SIMPLE_INTERRUPT() macro for USART1,2, and 3, yet our interrupt-driven occam-pi code is "just working."

To be investigated. Committing arch-specific code next.