Compressed io.module by creating arrays of register constants and indexing into them by USART number.
Modified wiring.module to use renamed PROCs in io.module. I'm not happy with the naming yet.

Added interrupt constants for the ATmega1280. For constant enumeration, I used the location of the vector in memory, so that there is some consistency between the numbering in our constants list and the manual.