Some changes to the build.sh script and Makefile.
build.sh now builds four firmwares: two for the 328p (at 8MHz and 16MHz) and two for the 1280 (at 8MHz and 16MHz). Passing --with-mcu or --with-fcpu to this script no longer has any meaning. Passing --prefix could be useful, though...

I made a few changes to the Makefile, mostly changing things, learning things. If my changes seem pointless, that's because I'm learning. The biggest change is it spits out a firmware with the name tvm-avr-$(MCU)-$(F_CPU).hex. I also want 'make install' to put the firmware in a sensible place, which is why passing --prefix to build.sh matters.

I'm assuming that some kind of package-building-driver will be calling these tools, so that passing --prefix will be handled by some other script. I've started reading the O'Reilly GNU Make book, and am slowly learning things... but I'm not yet at the point that I know how to integrate this into the build system directly.