Trying to break out host specific stuff sensibly.
For now it is only the detection of serial ports which is done, though the idea
is that other things which are platform specific will also have a hook in here.
This is all to try and make it easier to support multiple hosts without the
nasty if(osx) elif (linux) elif ... stuff that got littered in the code before.