This used to be "io.module" and "avr.module".
The intention is that you include "avr.module", which will include
the relevant modules for the architecture. In this case, we include
"base", "constants", "ffi", "usart", and "adc". 

We have not tested this code yet, but will do so shortly. We are reorganizing the two trees, and then will check the result in a package build.