First round of cleanups I plan to undertake. In plumbing.module, switched LEVEL from a BOOL to an INT, and cleaned up the serial code a good bit.
The switch from a BOOL to an INT should make the redundant declaration of HIGH and LOW between plumbing.module and wiring.module a non-issue. The cleaned up serial code should make it possible and easy to make read and write calls from plumbing.

I'm not nearly %100 on having both serial.read (which defaults to port 0) and serial.rx (which includes an argument for the port), but it seemed like a way to go for the time being. Any feedback would be appreciated.