This is a new building script.
It isn't, perhaps, very Python-esque. It is a translation of my 
bash script into Python. The option parsing is a bit more robust,
and there are more helper functions. My hope is that it will be
more maintainable than my bash-script hack.

There are usage instructions at the top of the file, or you can
run

python make-deb-pkg.py --help

and see if you can figure out what to do from there.