We're continuing our cleanup/renaming/reorganizing of plumbing.module.
Here, we've made a distinction between 

pin.mode, pin.read, pin.write

and

digital.mode, digital.read, digital.write

Specifically, the first is an interface in terms of registers, and the second in terms of a pin number. This way, we can re-use the pin.* PROCs in a more platform-neutral way. (We believe you should never have a ports[] reference in Plumbing, as it needs to be platform-agnostic.) Now, pin.location (soon to be renamed) becomes our platform-specific binding point, as opposed to *.mode/read/write.

(We think this makes more sense as we step away from the assumptions that 1. Plumbing should be built on Wiring, and 2. that the Arduino is the only board we're developing fore.)