These are now gross tweaks. Bad me.
I'm making it possible to refresh the package from a local repository, instead of having to check in/out. Mostly to speed development.

This is OK, as Adam has pointed out we can do this properly if we're willing to read directions. As if I have time to read directions. ;)

Either way, next commit will be meaningful.