For want of somewhere else...
Really, we should have a hardware test suite. This isn't that. But, perhaps... well, Adam mentioned this once: a "test shield" that you can plug in, run a suite of software, and find out if things are working.

I'll be committing small bits of code here to make sure that things work on the Arduino and Mega as we update all the innards of the arch and platforms code.