Removing the stand-alone pin list, as it was copied into the Scheme file.
The Scheme file should not need to be used ever again. However, it is how I generated the maps for the 1280, so I included it. My apologies for re-introducing Scheme into the tree... but this really was a *one-time* use.

chipmap.module contains a mapping for the digital pins on the Mega. This should be checked.

The Seeed Studio Mega will have a slightly different mapping... they expose more pins.