Working on a blocking read FFI call.
We need a way to quickly read in a buffer of text, and then pass that back to occam-pi. Specifically, the goal is to keep up with a GPS unit running at 9600.

Currently, this doesn't work... specifically, I think I have my array-of-function-pointers syntax wrong.

If there is a better/faster way to get data out of the serial port, do let me know. What bothers me is that this doesn't seem to be fast enough yet. (I'm attempting to move an 'if' out of the loop, which is why I'm reaching for the function pointers, so I can just look up the read_char() function once.)