[MATT] I thought I pushed this change at some point in the past. I have no idea why it isn't part of the tree.
The "#if defined()" statements that I introduced *must go*. This is a hack to eliminate some code that is 1280-specific. I'd rather it worked on all platforms, or it was handled better... either way, I'm not keen on the solution I introduced here.

This gets things building properly, which was the goal. A better solution needs to be found (separate email to developers@ on this subject will follow later.)