Adding in TWI code.
Last summer I wrote two TWI modules. Both are suspect, but at this point I know I need to have TWI in the tree permanently. The modules are not invasive, and therefore I don't mind bringing both in for the time being. 

I believe that twi2.module is the "better" of these two, as I learned from the writing of the first. If memory serves, both are incredibly slow.