Adding in TWI code.
This extends the interrupt table for the 328 and 1280 with the TWI vector, as well as providing the simple mapping. I know I tested this on the 328 when I was developing it, but in truth, this is a mildly adventuresome commit.

I'm committing this to trunk because 1. I don't believe it will break anything, and 2. we need TWI support. Or, I do.