Arch files for the 644p.
This should be enough to bootstrap a 644p.

-- constants.module
These look good at a glance. There are two USARTs on the 644, so I took 
  the m1280 constants and cut them down. The interrupt table needs to match 
  the one in the runtime.

-- iomxx4.inc
This is the header file, converted, for the 644p series. We'll see how it goes.

-- pwm.arch.module
This is a mess. I would say PWM probably doesn't work on the 644p using this
  module. I don't need PWM on the 644P for what I'm doing right now, so I'm 
  leaving this broken at the moment. 

-- external.interrupts.module
This doesn't look any different from the 328 at the moment, meaning we 
  only handle two external interrupts.

-- chipmap.module
This looks right. Or, if you prefer, I believe I wrote the mapping correctly for
  the 40-pin PDIP package.

-- ffi.module
Unchanged from 328p.

-- adc.module
I made minor changes here, mostly to account for the fact that ADC0 is pin 40
  on the 644p and we count down from there. This is handled, and should work.

-- avr.module
Unchanged. It pulls in all of these files. (Well, I changed the .inc file
  inclusion from the m328p.)