Extending the table for the 1280.
Again, I'm not happy about the #if statements. Like all (bad) hacks, I'm not going to figure it out right now.

This would extend the interrupt vector table from INT0:1 to INT0:7 on the Mega. Note, though, that:

1. Wiring only supports five interrupts. I'm not sure, right now, why that is. I have to look at a schematic. Perhaps they don't have all of the requisite pins connected? Perhaps they're used for something else.

2. This isn't tested. However, we did not have any external interrupts for the 1280 before I extended the occam side a few minutes ago (set.interrupts was implemented as a SKIP), so for this to not be "right" at this stage is moot.

I'll probably clean this up Sunday. There are one or two more things I want to check on the 1280 before we do another build/release, as there were some comments on the user list about 1280 oddities. It would be nice if we could check out what was being reported, fix it, and push a new build.