Remove full upload capability.
Although silly, this eliminates uploading. However, the (poor) design of the upload function means it does a bunch of important things before issuing the call to 'scp'. So, I simply removed the call to scp.

Yes, yes, the whole thing is a hack. But it drives a full build of multiple back-ends, etc., in one or two commands, which works for me for the moment.