Fixed occbuild, tweaked configure.ac.
I needed to bring occbuild up to Python3. Have to make sure this doesn't break a linux build.

I commented out occamdoc from the build for the moment... I don't know what is failing, and it is not my straight-line problem right now.

This builds on a Mac (10.13.6, High Sierra)

"builds" does not mean "works."

But "builds" can be worked with.